requirements.txt
langchain-openai
langgraph
langchain-core
gitpython
python-dotenv
